backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
Pillow>=12.1,<13
rembg==2.0.72
onnxruntime==1.22.1
pytesseract==0.3.13
requests==2.32.4
python-slugify==8.0.4
cryptography>=45,<46
